requirements.txt
Flask>=2.0
SQLAlchemy>=1.4
flask_sqlalchemy>=3.0
requests>=2.25
python-dotenv>=1.0